Hi @intx4!

I'd like to propose the following changes to your Arkworks PR. The first one concerns the sample size in the benches. I've noticed that it's set to 10 by default. I assume you chose this value due to how slow verification is. However, the sample size for tree creation and (multi-)path proving can be increased by an order of magnitude, and every bench will take roughly the same time to execute. 

The second change is refactoring the function `prefix_encode_path` using chained iterators. This approach is a bit more readable and even slightly more efficient. For reference, I get the following benches for the original implementation on my ThinkPad 14s machine: 
```
Merkle Tree Generate Multi Proof (Leaves as [u8])
                        time:   [470.45 ms 471.29 ms 472.20 ms]
```
When using iterators, the performance seems to improve slightly:
```
Merkle Tree Generate Multi Proof (Leaves as [u8])
                        time:   [462.76 ms 463.66 ms 464.64 ms]
                        change: [-1.8909% -1.6180% -1.3512%] (p = 0.00 < 0.05)
                        Performance has improved.
```