@intx4 2024-03-25
hi @Cesar199999, thank you very much for your additions, it was indeed a good catch to use iterators :)